requirements.txt
requests
beautifulsoup4
apscheduler
python-dotenv
fastapi
uvicorn